requirements.txt
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
psycopg2==2.9.10
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
setuptools==75.6.0
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
wheel==0.45.1
